make TrackerSystematicMisalignments a one module
